move tests from TestOzoneRpcClientWithRatis to parent class
